[FIX] move project page handler to projects container
